+++ lib/src/client/options.dart
@@ -17,6 +17,7 @@ import 'dart:math';
 
 import '../shared/codec_registry.dart';
 import 'client_keepalive.dart';
+import 'proxy.dart';
 import 'transport/http2_credentials.dart';
 
 const defaultIdleTimeout = Duration(minutes: 5);
@@ -59,6 +60,7 @@ class ChannelOptions {
   final BackoffStrategy backoffStrategy;
   final String userAgent;
   final ClientKeepAliveOptions keepAlive;
+  final Proxy? proxy;
 
   const ChannelOptions({
     this.credentials = const ChannelCredentials.secure(),
@@ -69,5 +71,6 @@ class ChannelOptions {
     this.connectionTimeout = defaultConnectionTimeOut,
     this.codecRegistry,
     this.keepAlive = const ClientKeepAliveOptions(),
+    this.proxy,
   });
 }