+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 3.2.4-wip
+## 3.2.4
 
 * Forward internal `GrpcError` on when throwing while sending a request.
 * Add support for proxies, see [#33](https://github.com/grpc/grpc-dart/issues/33).